+++ R/qsys_sge.r
@@ -10,6 +10,7 @@ SGE = R6::R6Class("SGE",
         },
 
         submit_jobs = function(n_jobs, template=list(), log_worker=FALSE) {
+            template = utils::modifyList(SGE$defaults, template)
             template$n_jobs = n_jobs
             template$master = private$master
             private$job_id = template$job_name = paste0("cmq", self$id)
@@ -60,6 +61,8 @@ SGE$setup = function() {
     user_defaults = getOption("clustermq.defaults")
     if (!is.null(user_defaults))
         SGE$defaults = user_defaults
+    else
+        SGE$defaults = list()
 
     SGE
 }