+++ R/qsys_slurm.r
@@ -10,6 +10,7 @@ SLURM = R6::R6Class("SLURM",
         },
 
         submit_jobs = function(n_jobs, template=list(), log_worker=FALSE) {
+            template = utils::modifyList(SLURM$defaults, template)
             template$n_jobs = n_jobs
             template$master = private$master
             private$job_id = template$job_name = paste0("cmq", self$id)
@@ -60,6 +61,8 @@ SLURM$setup = function() {
     user_defaults = getOption("clustermq.defaults")
     if (!is.null(user_defaults))
         SLURM$defaults = user_defaults
+    else
+        SLURM$defaults = list()
 
     SLURM
 }