+++ R/work_chunk.r
@@ -7,16 +7,23 @@
 #' @param const_args   Constant arguments passed to each call
 #' @param rettype        Return type of function
 #' @param common_seed  A seed offset common to all function calls
+#' @param progress     Logical indicated whether to display a progress bar
 #' @return             A list of call results (or try-error if they failed)
-work_chunk = function(df, fun, const_args=list(), rettype="list", common_seed=NULL) {
+work_chunk = function(df, fun, const_args=list(), rettype="list",
+                      common_seed=NULL, progress=FALSE) {
     context = new.env()
     context$warnings = list()
     context$errors = list()
+    if (progress)
+        pb = progress::progress_bar$new(total = nrow(df),
+                                        format = "[:bar] :percent eta: :eta")
 
     fwrap = function(..., ` id `, ` seed `=NA) {
         chr_id = as.character(` id `)
         if (!is.na(` seed `))
             set.seed(` seed `)
+        if (progress)
+            pb$tick()
 
         withCallingHandlers(
             withRestarts(