+++ R/qsys_lsf.r
@@ -6,18 +6,19 @@ LSF = R6::R6Class("LSF",
 
     public = list(
         initialize = function(...) {
-            super$initialize(...)
+            super$initialize(...,
+                template=getOption("clustermq.template", system.file("LSF.tmpl")))
         },
 
         submit_jobs = function(n_jobs, template=list(), log_worker=FALSE) {
-            template = utils::modifyList(LSF$defaults, template)
+            template = utils::modifyList(private$defaults, template)
             template$n_jobs = n_jobs
             template$master = private$master
             private$job_id = template$job_name = paste0("cmq", self$id)
             if (log_worker)
                 template$log_file = paste0(template$job_name, ".log")
 
-            filled = infuser::infuse(LSF$template, template)
+            filled = infuser::infuse(private$template, template)
 
             success = system("bsub", input=filled, ignore.stdout=TRUE)
             if (success != 0) {
@@ -46,33 +47,3 @@ LSF = R6::R6Class("LSF",
         job_id = NULL
     )
 )
-
-# Static method, process scheduler options and return updated object
-LSF$setup = function() {
-    user_template = getOption("clustermq.template.lsf")
-    if (!is.null(user_template)) {
-        warning("scheduler-specific templates are deprecated; use clustermq.template instead")
-        LSF$template = readChar(user_template, file.info(user_template)$size)
-    }
-    user_template = getOption("clustermq.template")
-    if (!is.null(user_template))
-        LSF$template = readChar(user_template, file.info(user_template)$size)
-
-    user_defaults = getOption("clustermq.defaults")
-    if (!is.null(user_defaults))
-        LSF$defaults = user_defaults
-    else
-        LSF$defaults = list()
-
-    LSF
-}
-
-# Static method, overwritten in qsys w/ user option
-LSF$template = paste(sep="\n",
-    "#BSUB-J {{ job_name }}[1-{{ n_jobs }}]    # name of the job / array jobs",
-    "#BSUB-o {{ log_file | /dev/null }}        # stdout + stderr",
-    "#BSUB-M {{ memory | 4096 }}               # Memory requirements in Mbytes",
-    "#BSUB-R rusage[mem={{ memory | 4096  }}]  # Memory requirements in Mbytes",
-    "",
-    "ulimit -v $(( 1024 * {{ memory | 4096 }} ))",
-    "R --no-save --no-restore -e 'clustermq:::worker(\"{{ master }}\")'")