+++ R/qsys_slurm.r
@@ -6,18 +6,19 @@ SLURM = R6::R6Class("SLURM",
 
     public = list(
         initialize = function(...) {
-            super$initialize(...)
+            super$initialize(...,
+                template=getOption("clustermq.template", system.file("SLURM.tmpl")))
         },
 
         submit_jobs = function(n_jobs, template=list(), log_worker=FALSE) {
-            template = utils::modifyList(SLURM$defaults, template)
+            template = utils::modifyList(private$defaults, template)
             template$n_jobs = n_jobs
             template$master = private$master
             private$job_id = template$job_name = paste0("cmq", self$id)
             if (log_worker)
                 template$log_file = paste0(template$job_name, ".log")
 
-            filled = infuser::infuse(SLURM$template, template)
+            filled = infuser::infuse(private$template, template)
 
             success = system("sbatch", input=filled, ignore.stdout=TRUE)
             if (success != 0) {
@@ -42,39 +43,9 @@ SLURM = R6::R6Class("SLURM",
     ),
 
     private = list(
+        template = "",
+        defaults = list(),
         is_cleaned_up = FALSE,
         job_id = NULL
     )
 )
-
-# Static method, process scheduler options and return updated object
-SLURM$setup = function() {
-    user_template = getOption("clustermq.template.slurm")
-    if (!is.null(user_template)) {
-        warning("scheduler-specific templates are deprecated; use clustermq.template instead")
-        SLURM$template = readChar(user_template, file.info(user_template)$size)
-    }
-    user_template = getOption("clustermq.template")
-    if (!is.null(user_template))
-        SLURM$template = readChar(user_template, file.info(user_template)$size)
-
-    user_defaults = getOption("clustermq.defaults")
-    if (!is.null(user_defaults))
-        SLURM$defaults = user_defaults
-    else
-        SLURM$defaults = list()
-
-    SLURM
-}
-
-# Static method, overwritten in qsys w/ user option
-SLURM$template = paste(sep="\n",
-    "#!/bin/sh",
-    "#SBATCH --job-name={{ job_name }}",
-    "#SBATCH --output={{ log_file | /dev/null }}",
-    "#SBATCH --error={{ log_file | /dev/null }}",
-    "#SBATCH --mem-per-cpu={{ memory | 4096 }}",
-    "#SBATCH --array=1-{{ n_jobs }}",
-    "",
-    "ulimit -v $(( 1024 * {{ memory | 4096 }} ))",
-    "R --no-save --no-restore -e 'clustermq:::worker(\"{{ master }}\")'")