+++ R/zzz.r
@@ -9,7 +9,7 @@
 
     if (length(qsys_default) == 0) {
         qname = c("SLURM", "LSF", "SGE", "LOCAL")
-        exec = Sys.which(c("sbatch", "bsub", "qsh"))
+        exec = Sys.which(c("sbatch", "bsub", "qsub"))
         select = c(which(nchar(exec) > 0), 4)[1]
         qsys_default = qname[select]
     }