+++ CHANGES.md
@@ -2,6 +2,7 @@
   * Fix error for `qsys$reusable` when using `n_jobs=0`/local processing (#75)
   * Scheduler-specific templates are deprecated. Use `clustermq.template` instead
   * Allow option `clustermq.defaults` to fill default template values (#71)
+  * Errors in `purrr` processing are now shut down cleanly (#67)
 
 * 0.8.3
   * Support `rettype` for function calls where return type is known (#59)