+++ CHANGES.md
@@ -1,3 +1,6 @@
+* 0.8.5
+  * Moved templates to package inst/ directory (#85)
+
 * 0.8.4
   * Fix error for `qsys$reusable` when using `n_jobs=0`/local processing (#75)
   * Scheduler-specific templates are deprecated. Use `clustermq.template` instead